chore(RedditPixel): move sdk loader to different file
