+++ src/integrations/RedditPixel/browser.js
@@ -8,9 +8,9 @@
 /* eslint-disable vars-on-top */
 /* eslint-disable no-unused-expressions */
 import logger from '../../utils/logUtil';
-import { LOAD_ORIGIN } from '../../utils/ScriptLoader';
 import { NAME } from './constants';
 import { getHashFromArrayWithDuplicate, getEventMappingFromConfig } from '../../utils/commonUtils';
+import { loader } from './loader';
 
 class RedditPixel {
   constructor(config, analytics, destinationInfo) {
@@ -29,19 +29,7 @@ class RedditPixel {
   init() {
     logger.debug('===In init RedditPixel===');
 
-    !(function (w, d) {
-      if (!w.rdt) {
-        var p = (w.rdt = function () {
-          p.sendEvent ? p.sendEvent.apply(p, arguments) : p.callQueue.push(arguments);
-        });
-        p.callQueue = [];
-        var t = d.createElement('script');
-        (t.src = 'https://www.redditstatic.com/ads/pixel.js'), (t.async = !0);
-        t.setAttribute('data-loader', LOAD_ORIGIN);
-        var s = d.getElementsByTagName('script')[0];
-        s.parentNode.insertBefore(t, s);
-      }
-    })(window, document);
+    loader(window, document);
 
     window.rdt('init', this.advertiserId);
   }
